+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Include XML declaration header in exported UDDF files
+- Include `xmlns` namespace attribute in exported UDDF files
 
 ## 0.3.0 - 2026-01-11